Migrate set of TES criteria that reference projected end date fields to sentencing v2 (Recidiviz/recidiviz-data#36665)
## Description of the change

This PR is stacked on Recidiviz/recidiviz-data#35631 and does a few things:

1. Injects sentence v1 states into
`sentence_inferred_group_serving_period_projected_dates`
2. Creates a new parallel view `sentence_serving_period_projected_dates`
(determined to be necessary as I was validating migrated criteria).
Sentence sessions [view
diagram](https://docs.google.com/drawings/d/1lHkCz8VI7SXna2quN6fHvglXO0OOSeSL7zlEORK7aS8/edit?usp=sharing)
is updated with this addition.
3. Writes a simple test for this new view
4. Migrates 4 TES criteria to the new V2 schema. These 4 criteria
represent the set of TES criteria that reference an array in
`sentence_spans` to obtain a projected date associated with a sentence
or collection of sentences (see
[here](https://www.notion.so/recidiviz/d1a4da82396d45a7907243e718d8187c?v=960ddb57385944b78c0d0e4362200ee3)).
Note that there are additional criteria that either reference
sentence_spans or reference a projected date _not_ through sentence
spans, but this is the set of criteria that I thought made sense to
start with when proving out the use of these new views.

[This](https://colab.research.google.com/drive/11C2cOICm8G5WO5oeifewIWaDaFSaKisp?usp=sharing)
notebook documents my process of validating output following this
migration.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Advances Recidiviz/recidiviz-data#33748

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Justine Kunz <justine@recidiviz.org>
GitOrigin-RevId: 0e1e4e9406141cc6ae482cb308433508f5632fc6