[Admin Panel] Incorporate `blockedOn` into state user permissions table (Recidiviz/recidiviz-data#37460)
## Description of the change

This PR replaces use of the `blocked` field with the `blockedOn` field
in `StateUserPermissions`. Currently, the `StateUserPermissions` table
contains a 'Blocked' column which displays an 'Enable user' button for
users who are blocked.

This PR replaces the 'Blocked' column with a 'Block status' column.
Users who are already blocked continue to display the 'Enable user'
button. Users who have an upcoming block date display a button that says
'Remove upcoming block [upcoming block date]'. The column can be
filtered and sorted by active users, users with upcoming blocks, and
blocked users.

I tested this locally by manually editing `blocked_on` in my local db to
be a future date.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#36604

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d2e37d23fe9d145b3a745e23ad4d8ed8ef9c8616