[TES] Enforce criteria conditions match top-level criteria (Recidiviz/recidiviz-data#38022)
## Description of the change

Add a verification step in the task eligibility spans view builder to
make sure that the criteria listed in the almost eligible condition
match a top-level criteria from the criteria builder list. Otherwise,
the query that constructs the eligibility spans view will not properly
incorporate the almost eligible logic. I've also added a test to make
sure the corresponding error is raised for this scenario.

This verification did flag the AZ overdue for Recidiviz DTP date
opportunity, and so I modified that to use the same
`OrTaskCriteriaGroup` in both lists. This updated the eligibility status
for 87 residents:
- 77 moved from almost eligible for TPR to almost eligible for DTP
- 10 were added as almost eligible for DTP

## Related issues

Closes Recidiviz/recidiviz-data#38014

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1eed0788652ca55df2e05f71051c707d90a1c517